# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /api/latest/hr-users/{username}:
    get:
      tags:
      - hr-controller
      operationId: getUser
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
    put:
      tags:
      - hr-controller
      operationId: updateUser
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
...
  /api/latest/user-info/locale:
    post:
      tags:
      - user-info-controller
      operationId: setLocale
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetLocaleRequestDto"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserInfoDto"
  /api/latest/user-info/coach:
    post:
      tags:
      - user-info-controller
      operationId: setCoach
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCoachRequestDto"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserInfoDto"
...
  /api/latest/admin/users:
    get:
      tags:
      - admin-view-controller
      operationId: getUsers
      parameters:
      - name: filterDto
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/FilterDto"
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageAdminView"
    post:
      tags:
      - admin-view-controller
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequestDto"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
  /api/latest/admin/users/{username}:
    post:
      tags:
      - admin-view-controller
      operationId: updateUser_1
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequestDto"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
    delete:
      tags:
      - admin-view-controller
      operationId: deleteUser
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
  /api/latest/admin/users/{username}/resent-invitation:
    post:
      tags:
      - admin-view-controller
      operationId: resentInvitation
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResentInvitationRequestDto"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
...
components:
  schemas:
    ErrorCodeFieldDto:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
    ErrorDto:
      type: object
      properties:
        errorCode:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/ErrorCodeFieldDto"
        errorMessage:
          type: string
...
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        username:
          type: string
        status:
          type: string
          enum:
          - AUTHORIZED
          - PENDING
          - PAID
          - REQUESTED
          - VIEWED
          - SUBMITTED
          - CANCELED
        timeZone:
          type: string
          minLength: 1
        authorities:
          type: array
          items:
            type: string
            enum:
            - RESPONDENT
            - USER
            - MANAGER
            - COACH
            - HR
            - ADMIN
          minItems: 1
          uniqueItems: true
        locale:
          type: string
          pattern: "[a-z]{2}"
        manager:
          type: string
        isManager:
          type: boolean
        position:
          type: string
        aspiredCompetency:
          type: string
      required:
      - authorities
      - firstName
      - lastName
      - status
      - timeZone
    UserDto:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        timeZone:
          type: string
        status:
          type: string
          enum:
          - AUTHORIZED
          - PENDING
          - PAID
          - REQUESTED
          - VIEWED
          - SUBMITTED
          - CANCELED
        authorities:
          type: array
          items:
            type: string
            enum:
            - RESPONDENT
            - USER
            - MANAGER
            - COACH
            - HR
            - ADMIN
          uniqueItems: true
        position:
          type: string
        manager:
          type: string
        isManager:
          type: boolean
        aspiredCompetency:
          type: string
...
    UserInfoDto:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        userRoles:
          type: array
          items:
            type: string
            enum:
            - RESPONDENT
            - USER
            - MANAGER
            - COACH
            - HR
            - ADMIN
          uniqueItems: true
        timeZone:
          type: string
        strengths:
          type: array
          items:
            type: string
        values:
          type: array
          items:
            type: string
        areaOfDevelopment:
          type: array
          items:
            type: string
        coach:
          type: string
        locale:
          type: string
        allowedCoachRates:
          type: array
          items:
            type: string
          uniqueItems: true
        companyId:
          type: integer
          format: int64
...
    SetLocaleRequestDto:
      type: object
      properties:
        locale:
          type: string
          pattern: "[a-z]{2}"
    SetCoachRequestDto:
      type: object
      properties:
        coach:
          type: string
...
    PageableObject:
      type: object
      properties:
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    CreateUserRequestDto:
      type: object
      properties:
        username:
          type: string
          format: email
          minLength: 1
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        timeZone:
          type: string
          minLength: 1
        companyId:
          type: integer
          format: int64
        isTrial:
          type: boolean
        authorities:
          type: array
          items:
            type: string
            enum:
            - RESPONDENT
            - USER
            - MANAGER
            - COACH
            - HR
            - ADMIN
          uniqueItems: true
        status:
          type: string
          enum:
          - AUTHORIZED
          - PENDING
          - PAID
          - REQUESTED
          - VIEWED
          - SUBMITTED
          - CANCELED
        locale:
          type: string
          pattern: "[a-z]{2}"
        rate:
          type: string
        internalRate:
          type: integer
          format: int32
        certificate:
          type: array
          items:
            type: string
          uniqueItems: true
      required:
      - firstName
      - isTrial
      - lastName
      - status
      - timeZone
      - username
    UpdateUserRequestDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        timeZone:
          type: string
        companyId:
          type: integer
          format: int64
        isTrial:
          type: boolean
        authorities:
          type: array
          items:
            type: string
            enum:
            - RESPONDENT
            - USER
            - MANAGER
            - COACH
            - HR
            - ADMIN
          uniqueItems: true
        status:
          type: string
          enum:
          - AUTHORIZED
          - PENDING
          - PAID
          - REQUESTED
          - VIEWED
          - SUBMITTED
          - CANCELED
        coach:
          type: string
        credit:
          type: integer
          format: int32
        freeCoach:
          type: string
        locale:
          type: string
          pattern: "[a-z]{2}"
        allowedCoachRates:
          type: array
          items:
            type: string
          uniqueItems: true
        rate:
          type: string
        internalRate:
          type: integer
          format: int32
        certificate:
          type: array
          items:
            type: string
          uniqueItems: true
    ResentInvitationRequestDto:
      type: object
      properties:
        locale:
          type: string
          pattern: "[a-z]{2}"
...
    FilterDto:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        timeZone:
          type: string
        status:
          type: string
          enum:
          - AUTHORIZED
          - PENDING
          - PAID
          - REQUESTED
          - VIEWED
          - SUBMITTED
          - CANCELED
        companyId:
          type: integer
          format: int64
        companyName:
          type: string
        coach:
          type: string
        coachFirstName:
          type: string
        coachLastName:
          type: string
        credit:
          type: integer
          format: int32
        requestedCredit:
          type: integer
          format: int32
        sumRequestedCredit:
          type: integer
          format: int32
        paidCredit:
          type: integer
          format: int32
        hrs:
          type: string
        requestedBy:
          type: string
        isTrial:
          type: boolean
        freeCoach:
          type: string
        maxCoachRate:
          type: string
        showCanceled:
          type: boolean
    AdminView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        authorities:
          type: array
          items:
            type: string
            enum:
            - RESPONDENT
            - USER
            - MANAGER
            - COACH
            - HR
            - ADMIN
          uniqueItems: true
        timeZone:
          type: string
        status:
          type: string
          enum:
          - AUTHORIZED
          - PENDING
          - PAID
          - REQUESTED
          - VIEWED
          - SUBMITTED
          - CANCELED
        companyId:
          type: integer
          format: int64
        companyName:
          type: string
        coach:
          type: string
        coachFirstName:
          type: string
        coachLastName:
          type: string
        credit:
          type: integer
          format: int32
        requestedCredit:
          type: integer
          format: int32
        sumRequestedCredit:
          type: integer
          format: int32
        paidCredit:
          type: integer
          format: int32
        scheduledCredit:
          type: integer
          format: int32
        hrs:
          type: string
        requestedBy:
          type: string
        freeCoach:
          type: string
        locale:
          type: string
        allowedCoachRates:
          type: string
        rate:
          type: string
        certificate:
          type: array
          items:
            type: string
          uniqueItems: true
        internalRate:
          type: integer
          format: int32
        primaryRoles:
          type: array
          items:
            type: string
            enum:
            - COACH
            - MENTOR
            - TRAINER
            - FACILITATOR
            - CONSULTANT
            - SPEAKER
          uniqueItems: true
    PageAdminView:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/AdminView"
        number:
          type: integer
          format: int32
        empty:
          type: boolean